requirements.txt
python-gnupg==0.3.8
PyYAML==3.11
requests==2.9.1
pybuilder==0.11.8
